[BE] [#214] Service 구조 변경(#216)
