+++ back-end/src/main/java/com/techie/backend/board/controller/PostController.java
@@ -3,7 +3,7 @@ package com.techie.backend.board.controller;
 import com.techie.backend.board.domain.PostCategory;
 import com.techie.backend.board.dto.PostRequest;
 import com.techie.backend.board.dto.PostResponse;
-import com.techie.backend.board.service.PostServiceImpl;
+import com.techie.backend.board.service.PostService;
 import com.techie.backend.global.security.UserDetailsCustom;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,33 +20,33 @@ import org.springframework.web.bind.annotation.*;
 @Slf4j
 public class PostController {
 
-    private final PostServiceImpl postServiceImpl;
+    private final PostService postService;
 
     @PostMapping
     public ResponseEntity<PostResponse> createPost(@RequestBody PostRequest postRequest,
                                                    @AuthenticationPrincipal UserDetailsCustom userDetails) {
-        PostResponse postResponse = postServiceImpl.createPost(postRequest, userDetails);
+        PostResponse postResponse = postService.createPost(postRequest, userDetails);
         return ResponseEntity.ok(postResponse);
     }
 
     @GetMapping("/my")
     public ResponseEntity<Page<PostResponse>> getMyPosts(@AuthenticationPrincipal UserDetailsCustom userDetails,
                                                          Pageable pageable) {
-        Page<PostResponse> myPosts = postServiceImpl.getMyPosts(userDetails, pageable);
+        Page<PostResponse> myPosts = postService.getMyPosts(userDetails, pageable);
         return ResponseEntity.ok(myPosts);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Page<PostResponse>> getYourPosts(@PathVariable Long id,
                                                            Pageable pageable) {
-        Page<PostResponse> yourPosts = postServiceImpl.getYourPosts(id, pageable);
+        Page<PostResponse> yourPosts = postService.getYourPosts(id, pageable);
         return ResponseEntity.ok(yourPosts);
     }
 
     @GetMapping("category/{category}")
     public ResponseEntity<Page<PostResponse>> getPostsByCategory(@PathVariable PostCategory category,
                                                                  Pageable pageable) {
-        Page<PostResponse> posts = postServiceImpl.getAllPosts(category, pageable);
+        Page<PostResponse> posts = postService.getAllPosts(category, pageable);
         return ResponseEntity.ok(posts);
     }
 
@@ -54,7 +54,7 @@ public class PostController {
     public ResponseEntity<Page<PostResponse>> searchPost(@RequestParam PostCategory category,
                                                          @RequestParam String query,
                                                          Pageable pageable) {
-        Page<PostResponse> posts = postServiceImpl.searchPost(category, query, pageable);
+        Page<PostResponse> posts = postService.searchPost(category, query, pageable);
         return ResponseEntity.ok(posts);
     }
 
@@ -62,14 +62,14 @@ public class PostController {
     public ResponseEntity<PostResponse> updatePost(@RequestBody PostRequest.Update updateRequest,
                                                    @PathVariable Long id,
                                                    @AuthenticationPrincipal UserDetailsCustom userDetails) {
-        PostResponse postResponse = postServiceImpl.updatePost(id, updateRequest, userDetails);
+        PostResponse postResponse = postService.updatePost(id, updateRequest, userDetails);
         return ResponseEntity.ok(postResponse);
     }
 
     @DeleteMapping
     public ResponseEntity<Void> deletePost(@RequestBody PostRequest.Delete delRequest,
                                            @AuthenticationPrincipal UserDetailsCustom userDetails) {
-        postServiceImpl.deletePost(delRequest, userDetails);
+        postService.deletePost(delRequest, userDetails);
         return ResponseEntity.noContent().build();
     }
 }
